+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Fixed
+- Added missing HTTP status codes
+
 ## 0.6.2 - 2016-10-17
 ### Changed
 - View engines now have HTML escaping enabled by default